requirements.txt
playwright>=1.58,<2.0
pytest>=9.0,<10.0
python-dotenv>=1.2,<2.0
